+++ SQLiteStudio3/guiSQLiteStudio/uiutils.h
@@ -14,6 +14,7 @@ GUI_API_EXPORT void setValidStateWihtTooltip(QWidget* widget, const QString& too
 GUI_API_EXPORT void setValidStateWarning(QWidget* widget, const QString& warning);
 GUI_API_EXPORT void setValidStateInfo(QWidget* widget, const QString& info);
 GUI_API_EXPORT void setValidStateTooltip(QWidget* widget, const QString& tip);
+GUI_API_EXPORT bool isValidStateIndicatorVisible(QWidget* widget);
 GUI_API_EXPORT const QStringList& getAllPageSizes();
 GUI_API_EXPORT QString convertPageSize(QPageSize::PageSizeId size);
 GUI_API_EXPORT QPageSize convertPageSize(const QString& size);